refactor(ew): reorganize post-related components and update routing
- Moved post-related components from the admin directory to a dedicated posts directory for better organization.
- Updated routing paths to reflect the new structure, ensuring all links point to the correct locations.
- Enhanced the CreatePostModal and related components to streamline post creation and management.
- Added new pages for viewing and editing posts, improving overall navigation and user experience.
