+++ apps/epic-web/src/app/dashboard/tips/page.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react'
+import { revalidateTag } from 'next/cache'
 import Link from 'next/link'
 import { redirect } from 'next/navigation'
 import type { Post } from '@/lib/posts'
@@ -7,7 +8,7 @@ import { getServerAuthSession } from '@/server/auth'
 import { cn } from '@/utils/cn'
 import { Lightbulb } from 'lucide-react'
 
-import { CreatePostModal } from '../../admin/posts/_components/create-post-modal'
+import { CreatePostModal } from '../../posts/_components/create-post-modal'
 
 /**
  * Page for a contributor to see and manage their tips.
@@ -58,7 +59,7 @@ const TipTeaser: React.FC<{
 	return (
 		<li className={cn('flex w-full items-center py-4', className)}>
 			<Link
-				href={`/admin/posts/${tip.fields.slug}/edit`}
+				href={`/posts/${tip.fields.slug}/edit`}
 				passHref
 				className="fluid-lg flex w-full items-center gap-3 py-5"
 			>
