+++ apps/epic-web/src/app/posts/_components/post-form-config.ts
