+++ apps/epic-web/src/app/posts/_components/post-player.tsx
