+++ apps/epic-web/src/app/posts/page.tsx
@@ -52,7 +52,7 @@ const PostTeaser: React.FC<{
 	return (
 		<li className={cn('', className)}>
 			<Link
-				href={`/admin/posts/${article.fields.slug}/edit`}
+				href={`/posts/${article.fields.slug}/edit`}
 				passHref
 				className="fluid-lg flex w-full items-center gap-3 py-5"
 			>
