+++ apps/epic-web/src/components/content/content-video-resource-field.tsx
@@ -6,7 +6,7 @@ import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { LessonPlayer } from '@/app/(content)/_components/lesson-player'
 import { NewLessonVideoForm } from '@/app/(content)/_components/new-lesson-video-form'
-import { SimplePostPlayer } from '@/app/admin/posts/_components/post-player'
+import { SimplePostPlayer } from '@/app/posts/_components/post-player'
 import Spinner from '@/components/spinner'
 import { env } from '@/env.mjs'
 import { useTranscript } from '@/hooks/use-transcript'
